Refactored code, added Docstring, and added TODO tasks
Added doc string and tidy up the code.  In the process
of filling in the docstring (following this guide
https://www.haskell.org/haddock/doc/html/markup.html)

Also added a few commands when you run "main"
When prompt to enter text, you can try:
  "What?"  or "What is this?"
  "Who?"  or "Who are we?"
  "Welcome"
  "Time?" or "What time it is?"  (hard coded the time)
  "exit"

Updated copyright holder name:

Updated the ArduinoSketch_MatrixLED.ino sketch to
display the Haskell Logo and the word "LED" on initialized.

Added TODO tasks in comments