Added new method 'clock' to the LEDDisplay Class
Added some HUnit tests

Added TODO for implementing the clock method for the LEDDisplay instance for LEDDisplaySettings
Renamed the concatMatrix to concatMatrixWithSeparator to be clear that it is not just
a simple concat of two matrix.
Renamed runForever to promptForever to be more accurate of its behavior
Added padFrameLeft, padFrameRight, and padFrameCenter to be able to render small frame.