Fixed the order of the matrix layout
The registers are wired from left to right (r1, r2, r3, r4), then the right most byte (8 LEDs) should be shift to r4, and the next right most byte should shift to r3 and eventually the last byte shift to r1.  The r0 is the layer control, so it will always be at r0.